update time formatting
